Выполнить работы по ТЗ 8
