+++ resources/task.csv
@@ -1,2 +1,10 @@
 id,type,name,status,description, startTime, duration, epic
+1,TASK,Составить список дел,NEW,Первая задача с изменениями,20.01.2025 08:30,30
+2,TASK,Скорректировать список дел,NEW,Вторая задача,20.02.2025 08:30,35
+3,EPIC,Спринт 4 с изменениями,NEW,Описание спринта +,20.01.2025 16:30,60
+6,EPIC,Переезд,NEW,Описание переезда,20.01.2026 08:30,30
+4,SUBTASK,Написать код,NEW,Написать код подробно,,100,3
+5,SUBTASK,Протестировать код,NEW,Протестировать код подробно,20.01.2025 16:30,60,3
+7,SUBTASK,Собрать коробки,NEW,Собрать все коробки,20.01.2026 08:30,30,6
 
+1,3,6,
