+++ test/FileBackedTaskManagerTest.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class FileBackedTaskManagerTest {
+class FileBackedTaskManagerTest extends TaskManagerTest<FileBackedTaskManager> {
 
     @Test
     void loadFromEmptyFile() throws IOException {
@@ -23,11 +23,9 @@ class FileBackedTaskManagerTest {
     @Test
     public void testException() {
         assertThrows(ManagerSaveException.class, () -> {
-            File file = File.createTempFile("testEmptyFile-", ".csv");
-            FileBackedTaskManager taskManager = new FileBackedTaskManager(file);
-            BufferedWriter writer = new BufferedWriter(new FileWriter(file));
+            FileBackedTaskManager taskManager = new FileBackedTaskManager(new File("resources/task*.csv"));
             taskManager.save();
-        }, "Попытка сохранить файл в который идет запись должна приводить к ошибке");
+        }, "Попытка сохранить файл должна приводить к ошибке");
     }
 
     @Test
@@ -45,7 +43,7 @@ class FileBackedTaskManagerTest {
     void saveLoadFile() throws IOException {
         File file = File.createTempFile("testEmptyFile-", ".csv");
         FileBackedTaskManager taskManager = new FileBackedTaskManager(file);
-        Task task = new Task("Test saveLoadFile", "Test saveLoadFile description",  TaskStatus.NEW, LocalDateTime.of(2025, 1,20,8,30), Duration.ofMinutes(30));
+        Task task = new Task("Test saveLoadFile", "Test saveLoadFile description", TaskStatus.NEW, LocalDateTime.of(2025, 1, 20, 8, 30), Duration.ofMinutes(30));
         final int taskId = taskManager.makeNewTask(task);
         //вызовем получение задачи для обновления истории
         taskManager.getTask(taskId);
@@ -54,5 +52,15 @@ class FileBackedTaskManagerTest {
         assertNotNull(taskManagerFromFile, "taskManagerFromFile is null!");
         assertEquals(taskManagerFromFile.getAllTasks().size(), taskManager.getAllTasks().size(), "Количество задач в менеджерах не равно");
         assertEquals(taskManagerFromFile.getHistory().size(), taskManagerFromFile.getHistory().size(), "Количество задач в истории менеджеров не равно");
+    }
+
+    @Override
+    TaskManager initTaskManager() {
+        try {
+            File file = File.createTempFile("testEmptyFile-", ".csv");
+            return  FileBackedTaskManager.loadFromFile(file);
+        } catch (IOException ignored) {
+        }
+        return null;
     }
 }
