+++ test/InMemoryTaskManagerTest.java
@@ -1,0 +1,10 @@
+import java.io.File;
+import java.io.IOException;
+
+public class InMemoryTaskManagerTest extends TaskManagerTest<InMemoryTaskManager> {
+
+    @Override
+    TaskManager initTaskManager() {
+        return new InMemoryTaskManager();
+    }
+}
